Add lint step to ci (#67)
The main objective with this pr is to add a lint step to our ci
pipeline.

When running this on our repository we get a lot of deprecation issues.
Some are trivial to fix e.g. `NamespaceFormatter`.

In https://github.com/grafana/authlib/pull/63 we moved them to claims
package but did not replace internal usage. This will break existing
code but should be easy to fix by just importing the correct formatter
from claims package instead.

There are other changes we probably need to do such as replacing usage
of `authn.CallerAuthInfo` inside our authz client code with the
interface `claims.AuthInfo`. Those are a bit bigger changes and I prefer
to do them in a separate pr.