Docs: Clean up and latest documentation (#91)
This PR focuses on improving the clarity and usability of the
documentation:

* **Better Organization**: Reorganized content to improve flow and
readability. Adding more info about the authn and authz specifics in the
dedicated packages. Linking to the RBAC plugin example for more details
on the single tenant usage of the library.
* **Code Examples**: Added code snippets with context, explaining the
purpose of the various library elements, focusing a bit on gRPC
interceptors.
* **Authentication / Authorization flows**: Explained the various
authentication and authorization flows we have in our Grafana ecosystem.

---------

Co-authored-by: Claudiu Dragalina-Paraipan <drclau@users.noreply.github.com>