Pass authz's http client to verifier (#70)
The PR passes the HTTP client of `authz` to the `KeyRetriever`. This
ensures any options passed to `authz` client will be passed to
`KeyRetriever`. A common use case is to skip TLS checks in development
environment where this option must be passed to `KeyRetriever` as well
to ignore certificate errors while fetching signing keys.

Signed-off-by: Mahendra Paipuri <mahendra.paipuri@gmail.com>