AuthZ: Multi-tenant client only requires one parent (#99)
This PR refactors the Multi-Tenant AccessControl client to use a new
`check` endpoint instead of the `read` endpoint. This will allow us not
to require the entire list of parents of a given resource. The AuthZ
Service (then Grafana) will handle the expansion of the folder tree.

**Changes**
- Migrated the client to use the `/check ` endpoint of the Authz
service.
- Updated the `CheckRequest` struct to be more aligned with k8s:
`namespace`, `resource`, `attribute`, `name`, and `parent` fields.

**Benefits:**
- The client will be simpler to use

**Next steps:**
- Update the documentation to reflect the changes to the RBAC system.
- Implement the server side of this change.