Folders: Hide k6 folders for users and show them for SAs (#143)
**What**

Don't allow users to see or list k6 folders, but allow service accounts
to access them.

This mirrors the work that we did previously to prevent users from
accidentally deleting k6 folders, which are used to scope k6 projects
and their permissions.

Previous PRs on the legacy code paths:
https://github.com/grafana/grafana/pull/88884 and
https://github.com/grafana/grafana/pull/90008

Note: I did initially want to do this in the code that lists folders,
but it required too many changes to the search code, so this is simpler
and probably also safer.