Chore: cleanup authenticator and client interceptor (#149)
I looked through the usage of client interceptor within grafana.

We no longer use `WithDisableAccessTokenOption` so that was removed. We
also never use any of these options:
* `WithMetadataExtractorOption`
* `AccessTokenMetadataKey`
* `IDTokenMetadataKey`

So to clean this component up a bit I changed it so that this one
accepts a `TokenExchanger` and move everything else to options.

Additional changes
* Rename all options to have matching prefix `WithClientInterceptor`
* Added `StaticTokenExchanger` that we can use instead of
https://github.com/grafana/grafana/blob/b850c9fa68930bd94d93caae524f2e3dabf79218/pkg/services/authn/grpcutils/inproc_exchanger.go#L14C6-L14C21.
* Removed `IDTokenAuthenticator`, we did not end up using this one