Authz: namespace validation (#153)
In https://github.com/grafana/grafana/pull/101124 we discovered that it
is impossible to use watch across namespaces and this is considered a
bug.

Watching for a resource type across namespaces will happen whenever we
deploy multi-tenant operators.

I changed the validation checks to never check for empty namespaces and
moved the namespace matches check into validation instead so its done
both for `FixedAccessClient` and authz implementation. This is safe
because only `TypeAccessPolicy` will ever be able to perform this kind
of watch. Any other type is forwarded to authz service where an empty
namespace woul always be considered errors. Also there _should_ never be
a case where we get an identity without the namespace set for it so it
_should_ never match an empty namespace. It is also only possible for
`TypeAccessPolicy` to have wildcard namespace and that is the only way
to match it.

Other change is this pr:
I removed namespace from access checker, empty namespace will only work
for access policies and for them we use permissions from the token, all
request that ends up doing a list to authz-service will only return
items from one namespace.

Fixes https://github.com/grafana/authlib/issues/152