Docs: Update documentation (#157)
Instead of having docs in REAME I create a docs folder. In this folder I
added `authorization` and `authentication` where I put what we currently
have. We should extend the docs with examples for other scenarios but I
will leave that for another pr.

I remove the example using `inproc` because at least right now that is
highly specific to grafana.

---------

Co-authored-by: Ieva <vasiljeva.ieva@gmail.com>