identity: Add support for custom expiration when exchanging tokens (#169)
Part of https://github.com/grafana/identity-access-team/issues/1217.
Relates to https://github.com/grafana/auth/pull/854.

---

This PR introduces support for specifying a custom expiration when
exchanging tokens. If not provided, the default will be 10 minutes.

As a side note, if the signing access policy token defines a
`maxExpiration` attribute, a 400-level error will be returned by the
API.