Add WithCache option to for NewKeyRetriever (#166)
It might be worth offering the option to set a custom cache, possibly
with different policies or with a different caching store.

This is almost supported, as cache.Cache is an interface, so it's
possible to reimplement it. This is done using a
DefaultKeyRetrieverOption function to pass to NewKeyRetriever that can
replace the cache field `c`. The problem is that the field is not
exported, so it's not possible to implement the option outside of the
package.

Add that option here. It needs to take into consideration the fact that
cache.NewLocalCache ends up starting a goroutine to do clean up tasks.
While the goroutine should stop because the cache has been replaced,
it's better to not start it in the first place.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>