Implement retry when calling the TokenExchangeURL (#181)
This PR introduces a retry logic when calling the external Token
Exchange URL.

As of today, if the call fails because of a networking issue, the entire
Exchange step fails too.

The current PR adds backoff mechanism that will retry the call for 3
times, with a max backoff interval of 4 seconds.

Fixes: https://github.com/grafana/search-and-storage-team/issues/326

---------

Signed-off-by: Maicon Costa <maiconscosta@gmail.com>
Co-authored-by: Jean-Philippe Quéméner <jeanphilippe.quemener@grafana.com>