AccessChecker:  Move folder to top level parameter (#193)
The AccessChecker requires the caller to provide a folder when a
resource exists in a folder. This interface requires the caller to
provide the folder relationship explicitly.

This design choice was made so we can avoid having zanzana needing to
watch *every* resource and stay in sync, however it is very easy to miss
this giant gotcha when using the client.

Rather than passing in folder as one of many optional parameters, this
PR makes it an explicitly required top-level parameter